sourcecode/requirements.txt
absl-py==0.7.1
astor==0.8.0
bert-serving-client==1.9.8
bert-serving-server==1.9.8
blis==0.2.4
boto==2.49.0
boto3==1.9.188
botocore==1.12.188
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
colorama==0.4.1
cycler==0.10.0
cymem==2.0.2
Cython==0.29.13
decorator==4.4.0
docutils==0.14
en-core-web-sm==2.1.0
fairseq==0.7.1
gast==0.2.2
gensim==3.8.1
GMatch4py==0.2.5b0
google-pasta==0.1.7
GPUtil==1.4.0
grpcio==1.22.0
h5py==2.9.0
idna==2.8
jmespath==0.9.4
joblib==0.13.2
jsonlines==1.2.0
jsonschema==2.6.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.1
mock==3.0.5
murmurhash==1.0.2
networkx==2.3
neuralcoref==4.0
nltk==3.4.4
numpy==1.16.4
pandas==0.24.2
Pillow==6.1.0
plac==0.9.6
portalocker==1.5.0
preshed==2.0.1
protobuf==3.9.0
psutil==5.6.3
pycparser==2.19
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1
pyzmq==18.0.2
requests==2.22.0
s3transfer==0.2.1
sacrebleu==1.3.7
scikit-learn==0.21.2
scipy==1.3.0
six==1.12.0
sklearn==0.0
smart-open==1.8.4
spacy==2.1.0
srsly==0.0.7
tensorboard==1.13.1
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.1
termcolor==1.1.0
thinc==7.0.8
torch==1.1.0
torchvision==0.3.0
tqdm==4.32.2
typing==3.7.4
urllib3==1.25.3
wasabi==0.2.2
Werkzeug==0.15.4
wrapt==1.11.2
